updating my branch to lastest version

+++ src/SortProduct.java
@@ -3,13 +3,13 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 public class SortProduct{
-    public static ArrayList<Double> sortAscending(ArrayList<Double> prices){
+    public static ArrayList<Long> sortAscending(ArrayList<Long> prices) {
         Collections.sort(prices);
 
         return prices;
     }
 
-    public static ArrayList<Double> sortDescending(ArrayList<Double> prices){
+    public static ArrayList<Long> sortDescending(ArrayList<Long> prices) {
         prices.sort(Collections.reverseOrder());
 
         return prices;
